refactor some resolvers
