--- internal/resolvers/twitch/model.go
@@ -1,10 +1,0 @@
-package twitch
-
-type twitchClipsTooltipData struct {
-	Title        string
-	AuthorName   string
-	ChannelName  string
-	Duration     string
-	CreationDate string
-	Views        string
-}
